moved graph_operations test into aid_utils
